Move helper functions to usePool
